refactor: Refactor auth requests logic
